[step4] refactor: 불필요 패키지 & 파일 제거
